search bar in main toolbar
