+++ app/src/main/AndroidManifest.xml
@@ -12,12 +12,17 @@
         android:supportsRtl="true"
         android:theme="@style/InstagramSearchAppTheme">
 
-        <activity android:name=".view.MainActivity">
+        <activity android:name=".main.MainActivity"
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
+                <!--<action android:name="android.intent.action.SEARCH" />-->
             </intent-filter>
-        </activity>
 
+            <!--<meta-data-->
+                <!--android:name="android.app.searchable"-->
+                <!--android:resource="@xml/searchable" />-->
+        </activity>
     </application>
 </manifest>
