+++ app/src/main/res/layout/activity_main.xml
@@ -7,10 +7,11 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context="io.toru.instagramsearch.view.MainActivity">
+        tools:context="io.toru.instagramsearch.main.MainActivity">
 
         <android.support.v7.widget.Toolbar
             style="@style/BasicToolbarTheme"
+            android:id="@+id/toolbar_main"
             app:title="@string/app_name"/>
 
         <android.support.v7.widget.RecyclerView
