+++ app/src/main/res/menu/search_menu.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <!-- android:icon="@android:drawable/ic_menu_search" -->
+    <item
+        android:id="@+id/action_search"
+        android:icon="@drawable/ic_search_white_24px"
+        android:title="@string/search"
+        app:showAsAction="always|collapseActionView"
+        app:actionViewClass="android.support.v7.widget.SearchView"/>
+</menu>
