+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">InstagramSearch</string>
+    <string name="search">Search</string>
 </resources>
